fix: design system linting issues
