+++ clients/design-system/src/components/Button/Button.type.ts
@@ -1,0 +1,10 @@
+export enum ButtonSize {
+  'small' = 'small',
+  'medium' = 'medium',
+  'large' = 'large',
+}
+
+export enum ButtonVariant {
+  'default' = 'default',
+  'contained' = 'contained',
+}
