+++ clients/design-system/src/components/Hero/Hero.tsx
@@ -1,4 +1,4 @@
-import React, { Children } from 'react';
+import React from 'react';
 import './style.module.scss';
 
 export interface HeroProps {
