+++ clients/design-system/src/components/Typography/Typography.type.ts
@@ -1,0 +1,15 @@
+export enum TypographyVariants {
+  'h1' = 'h1',
+  'h2' = 'h2',
+  'h3' = 'h3',
+  'h4' = 'h4',
+  'h5' = 'h5',
+  'h6' = 'h6',
+  'body' = 'body',
+  'intro' = 'intro',
+}
+
+export enum TypographyAlign {
+  'center' = 'center',
+  'left' = 'left',
+}
